updates postinstall to be safer

+++ .wonderpress/postinstall.js
@@ -10,12 +10,15 @@ import fs from 'fs';
 console.log('Running various postinstall scripts...');
 
 // Create an archive directory for storing various things
-fs.mkdir('.wonderpress/archive', true, (err) => {
-    if (err) {
-        return console.error(err);
-    }
-    console.log('Directory created successfully!');
-});
+const archivePath = '.wonderpress/archive';
+if(!fs.existsSync(archivePath)) {
+	fs.mkdir(archivePath, true, (err) => {
+	    if (err) {
+	        return console.error(err);
+	    }
+	    console.log('Archive directory created successfully!');
+	});
+}
 
 // Copy the original package.json, in case we need to reset later
 fs.copyFile('package.json', '.wonderpress/archive/package.json', (err) => {
@@ -30,15 +33,58 @@ fs.copyFile('.gitignore', '.wonderpress/archive/.gitignore', (err) => {
 	console.log('Original .gitignore was copied to .wonderpress/archive/.gitignore');
 });
 
-// Install Static Kit
-execSync("npx statickit .");
 
-// We no longer need the Static Kit CLI, remove it
-execSync("npm uninstall -D @wndrfl/static-kit-cli");
-execSync("npm prune");
+// Attempt to install Static Kit
 
-// Copy the original .gitignore back
-fs.copyFile('.wonderpress/archive/.gitignore', '.gitignore', (err) => {
-	if (err) throw err;
-	console.log('Original .gitignore was restored');
+console.log('Attempting to detect if it is safe to install Static Kit...');
+
+let isSafeToInstall = true;
+
+const shouldNotExist = [
+	'statickit.json',
+	'css',
+	'js',
+	'images'
+];
+
+shouldNotExist.forEach((path) => {
+	if(fs.existsSync(path)) {
+		isSafeToInstall = false;
+	}
 });
+
+if(!isSafeToInstall) {
+	console.log('Detected files or folders that would be overwritten by installing Static Kit. Moving on...');
+
+} else {
+
+	console.log('It appears safe to install Static Kit. Proceeding...');
+	
+	// Install Static Kit
+	execSync("npx statickit .");
+
+	// We no longer need the Static Kit CLI, remove it
+	execSync("npm uninstall -D @wndrfl/static-kit-cli");
+	execSync("npm prune");// Copy the original .gitignore back
+	
+	fs.copyFile('.wonderpress/archive/.gitignore', '.gitignore', (err) => {
+		if (err) throw err;
+		console.log('Original .gitignore was restored');
+	});
+}
+
+
+// Remove postinstall script from the npm scripts
+// This will prevent later developers from accidentally installing
+// Static Kit again...
+
+// Grab the package.json
+let newPackageData = fs.readFileSync('package.json');
+let newPackageJson = JSON.parse(newPackageData);
+
+if(newPackageJson.scripts && newPackageJson.scripts.postinstall) {
+	delete newPackageJson.scripts.postinstall;
+}
+
+let data = JSON.stringify(newPackageJson, null, 2);
+fs.writeFileSync('package.json', data);
